+++ demo/backend/components/forms/checkboxes.py
@@ -1,14 +1,23 @@
 from django import forms
 from django.forms import MultipleHiddenInput
 
+from crispy_forms_gds.choices import Choice
 from crispy_forms_gds.helper import FormHelper
 from crispy_forms_gds.layout import HTML, Button, Field, Layout, Size
 
 
 class CheckboxesForm(forms.Form):
 
+    METHODS = (
+        Choice("email", "Email"),
+        Choice(
+            "phone", "Phone", hint="Select this option only if you have a mobile phone"
+        ),
+        Choice("text", "Text message"),
+    )
+
     method = forms.MultipleChoiceField(
-        choices=(("email", "Email"), ("phone", "Phone"), ("text", "Text message"),),
+        choices=METHODS,
         widget=forms.CheckboxSelectMultiple,
         label="How would you like to be contacted?",
         help_text="Select all options that are relevant to you.",
@@ -19,12 +28,7 @@ class CheckboxesForm(forms.Form):
         super().__init__(*args, **kwargs)
         self.helper = FormHelper()
         self.helper.layout = Layout(
-            Field.checkboxes(
-                "method",
-                legend_size=Size.MEDIUM,
-                small=True,
-                hints={"phone": "Select this option only if you have a mobile phone"},
-            ),
+            Field.checkboxes("method", legend_size=Size.MEDIUM, small=True),
             Button("submit", "Submit"),
         )