+++ src/crispy_forms_gds/layout/fields.py
@@ -82,7 +82,7 @@ class Field(crispy_forms_layout.LayoutObject):
 
     @classmethod
     def checkboxes(
-        cls, field, legend_size=None, legend_tag=None, small=False, hints=None, **kwargs
+        cls, field, legend_size=None, legend_tag=None, small=False, **kwargs
     ):
         """
         Create a field for displaying a set of checkboxes.
@@ -99,9 +99,6 @@ class Field(crispy_forms_layout.LayoutObject):
 
             small (bool): Display small checkboxes. Default is False.
 
-            hints (dict): Hints that will be displayed for one or more checkboxes.
-                The dictionary key is the checkbox value. Defaults to None.
-
             **kwargs: Attributes to add to the <input> element when the field is
                 rendered.
 
@@ -116,9 +113,6 @@ class Field(crispy_forms_layout.LayoutObject):
 
         context["checkboxes_small"] = small
 
-        if hints:
-            context["checkboxes_hints"] = hints
-
         return Field(field, context=context, **kwargs)
 
     @classmethod
@@ -129,7 +123,6 @@ class Field(crispy_forms_layout.LayoutObject):
         legend_tag=None,
         small=False,
         inline=False,
-        hints=None,
         **kwargs
     ):
         """
@@ -148,9 +141,6 @@ class Field(crispy_forms_layout.LayoutObject):
 
             inline (bool): Display the radio buttons in a row. Default is False.
 
-            hints (dict): Hints that will be displayed for one or more checkboxes.
-                The dictionary key is the checkbox value. Defaults to None.
-
             **kwargs: Attributes to add to the <input> element when the field is
                 rendered.
 
@@ -165,9 +155,6 @@ class Field(crispy_forms_layout.LayoutObject):
 
         context["radios_small"] = small
         context["radios_inline"] = inline
-
-        if hints:
-            context["radios_hints"] = hints
 
         return Field(field, context=context, **kwargs)